requirements.txt
pillow==10.4.0
psutil==6.0.0
pyusb==1.2.1
